Use 64-bit offsets when repairing PDF objects.

+++ source/pdf/pdf-repair.c
@@ -10,8 +10,8 @@ struct entry
 {
 	int num;
 	int gen;
-	int ofs;
-	int stm_ofs;
+	int64_t ofs;
+	int64_t stm_ofs;
 	int stm_len;
 };
